backend/requirements.txt
Flask
Flask-CORS
requests
tensorflow
transformers
spacy
firebase-admin
twilio
python-dotenv